imp: useBlockPromiseMultipleClick 훅 테스트코드 및 네이밍 수정 완료
